fsarchiver: provide spec for version 0.6.17
Provide spec file for building fsarchiver 0.6.17 on el6, i386 and x86_64.

FSArchiver is a system tool that allows you to save the contents of a
file-system to a compressed archive file. The file-system can be
restored  on a partition which has a different size and it can be
restored on a different file-system. Unlike tar/dar, FSArchiver also
creates the file-system when it extracts the data to partitions.
Everything is checksummed in the archive in order to protect the data.
If the archive is corrupt, you just lose the current file, not the whole
archive.